requirements.txt
pymupdf
rapidocr-onnxruntime
fastapi
uvicorn
python-multipart
numpy
python-docx
Pillow
openpyxl
python-pptx
pyjwt
passlib[bcrypt]
stripe
python-dotenv
python-magic
slowapi
